+++ packages/gits_http/lib/src/cache_strategy/model/cache_wrapper.dart
@@ -1,0 +1,53 @@
+import 'dart:convert';
+
+import 'package:http/http.dart';
+
+class CacheWrapper {
+  CacheWrapper({
+    required this.cacheDate,
+    required this.response,
+  });
+
+  final int cacheDate;
+  final Response response;
+
+  Map<String, dynamic> toMap() {
+    return {
+      'cacheDate': cacheDate,
+      'response': {
+        'body': response.body,
+        'headers': response.headers,
+        'is_redirect': response.isRedirect,
+        'persistent_connection': response.persistentConnection,
+        'reason_phrase': response.reasonPhrase,
+        'request': {
+          'url': response.request?.url.toString(),
+          'method': response.request?.method,
+        },
+      },
+    };
+  }
+
+  factory CacheWrapper.fromMap(Map<String, dynamic> map) {
+    return CacheWrapper(
+      cacheDate: map['cacheDate']?.toInt() ?? 0,
+      response: Response(
+        map['response']?['body'] ?? '',
+        map['response']?['status_code'] ?? '',
+        headers: map['response']?['headers'],
+        isRedirect: map['response']?['is_redirect'],
+        persistentConnection: map['response']?['persistent_connection'],
+        reasonPhrase: map['response']?['reason_phrase'],
+        request: Request(
+          map['response']?['request']?['method'] ?? '',
+          Uri.parse(map['response']?['request']?['url'] ?? ''),
+        ),
+      ),
+    );
+  }
+
+  String toJson() => json.encode(toMap());
+
+  factory CacheWrapper.fromJson(String source) =>
+      CacheWrapper.fromMap(json.decode(source));
+}