+++ packages/gits_http/lib/src/cache_strategy/strategy/just_async_strategy.dart
@@ -1,0 +1,14 @@
+import 'package:http/http.dart';
+
+import '../storage/storage.dart';
+import 'cache_strategy.dart';
+
+class JustAsyncStrategy extends CacheStrategy {
+  @override
+  Future<Response> applyStrategy(
+      {required String key,
+      required Storage storage,
+      required Future<Response> Function() fetch}) {
+    return invokeAsync(key: key, storage: storage, fetch: fetch);
+  }
+}