+++ packages/gits_http/lib/src/cache_strategy/strategy/just_cache_strategy.dart
@@ -1,0 +1,35 @@
+import 'package:http/http.dart';
+
+import '../../errors/gits_exceptions.dart';
+import '../storage/storage.dart';
+import 'cache_strategy.dart';
+
+class JustCacheStrategy extends CacheStrategy {
+  JustCacheStrategy({
+    this.ttlValue = CacheStrategy.defaultTTLValue,
+    this.keepExpiredCache = false,
+  });
+
+  final Duration ttlValue;
+  final bool keepExpiredCache;
+
+  @override
+  Future<Response> applyStrategy({
+    required String key,
+    required Storage storage,
+    required Future<Response> Function() fetch,
+  }) async {
+    final cached = await invokeCache(
+      key: key,
+      storage: storage,
+      ttlValue: ttlValue,
+      keepExpiredCache: keepExpiredCache,
+    );
+
+    if (cached == null) {
+      throw CacheException();
+    }
+
+    return cached;
+  }
+}