+++ packages/gits_http/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0-dev
+
+- add cache strategy (AsyncOrCacheStrategy, CacheOrAsyncStrategy, JustAsyncStrategy, JustCacheStrategy), default JustAsyncStrategy
+- remove head method in GitsHttp
+- remove read method in GitsHttp
+- remove readBytes method in GitsHttp
+
 ## 1.1.3
 
 - update dependency gits_inspector to 2.0.1